+++ KernelBench/level1/14_Matmul_for_upper_triangular_matrices.py
@@ -30,8 +30,8 @@ def get_inputs():
     Returns:
         list: A list containing two upper triangular matrices of shape (N, N).
     """
-    A = torch.triu(torch.rand(N, N))
-    B = torch.triu(torch.rand(N, N))
+    A = torch.triu(torch.rand(N, N, dtype=torch.float16))
+    B = torch.triu(torch.rand(N, N, dtype=torch.float16))
     return [A, B]
 
 def get_init_inputs():
