+++ KernelBench/level1/16_Matmul_with_transposed_A.py
@@ -26,8 +26,8 @@ K = 4096 * 2
 N = 2048 * 2
 
 def get_inputs():
-    A = torch.rand(K, M)
-    B = torch.rand(K, N)
+    A = torch.rand(K, M, dtype=torch.float16)
+    B = torch.rand(K, N, dtype=torch.float16)
     return [A, B]
 
 def get_init_inputs():
