+++ KernelBench/level1/31_ELU.py
@@ -32,7 +32,7 @@ batch_size = 4096
 dim = 393216
 
 def get_inputs():
-    x = torch.rand(batch_size, dim)
+    x = torch.rand(batch_size, dim, dtype=torch.float16)
     return [x]
 
 def get_init_inputs():
