+++ KernelBench/level1/41_Max_Pooling_1D.py
@@ -43,7 +43,7 @@ dilation    = 3
 return_indices = False
 
 def get_inputs():
-    x = torch.rand(batch_size, features, sequence_length)
+    x = torch.rand(batch_size, features, sequence_length, dtype=torch.float16)
     return [x]
 
 def get_init_inputs():
