+++ KernelBench/level1/42_Max_Pooling_2D.py
@@ -40,7 +40,7 @@ padding = 1
 dilation = 1
 
 def get_inputs():
-    x = torch.rand(batch_size, channels, height, width)
+    x = torch.rand(batch_size, channels, height, width, dtype=torch.float16)
     return [x]
 
 def get_init_inputs():
