+++ KernelBench/level1/52_Argmin_over_a_dimension.py
@@ -33,7 +33,7 @@ dim2 = 4095
 dim = 1
 
 def get_inputs():
-    x = torch.rand(batch_size, dim1, dim2)
+    x = torch.rand(batch_size, dim1, dim2, dtype=torch.float16)
     return [x]
 
 def get_init_inputs():
