+++ KernelBench/level1/90_cumprod.py
@@ -37,7 +37,7 @@ input_shape = (32768,)
 dim = 1
 
 def get_inputs():
-    return [torch.rand(batch_size, *input_shape)]
+    return [torch.rand(batch_size, *input_shape, dtype=torch.float32)]
 
 def get_init_inputs():
     return [dim]
