+++ KernelBench/level1/9_Tall_skinny_matrix_multiplication_.py
@@ -25,8 +25,8 @@ M = 16384 * 2
 N = 16 * 2
 
 def get_inputs():
-    A = torch.rand(M, N)
-    B = torch.rand(N, M)
+    A = torch.rand(M, N, dtype=torch.float16)
+    B = torch.rand(N, M, dtype=torch.float16)
     return [A, B]
 
 def get_init_inputs():
