+++ KernelBench/level2/39_Gemm_Scale_BatchNorm.py
@@ -8,7 +8,7 @@ class Model(nn.Module):
     def __init__(self, in_features, out_features, scale_shape, eps=1e-5, momentum=0.1):
         super(Model, self).__init__()
         self.gemm = nn.Linear(in_features, out_features)
-        self.scale = nn.Parameter(torch.randn(scale_shape))
+        self.scale = nn.Parameter(torch.randn(scale_shape, dtype=torch.float16))
         self.bn = nn.BatchNorm1d(out_features, eps=eps, momentum=momentum)
 
     def forward(self, x):
@@ -23,7 +23,7 @@ out_features = 4096
 scale_shape = (out_features,)
 
 def get_inputs():
-    return [torch.rand(batch_size, in_features)]
+    return [torch.rand(batch_size, in_features, dtype=torch.float16)]
 
 def get_init_inputs():
     return [in_features, out_features, scale_shape]
