+++ KernelBench/level2/51_Gemm_Subtract_GlobalAvgPool_LogSumExp_GELU_ResidualAdd.py
@@ -8,7 +8,7 @@ class Model(nn.Module):
     def __init__(self, in_features, out_features, bias=True):
         super(Model, self).__init__()
         self.gemm = nn.Linear(in_features, out_features, bias=bias)
-        self.subtract = nn.Parameter(torch.randn(out_features))
+        self.subtract = nn.Parameter(torch.randn(out_features, dtype=torch.float16))
 
     def forward(self, x):
         original_x = x.clone().detach()
@@ -37,7 +37,7 @@ in_features = 8192
 out_features = 8192
 
 def get_inputs():
-    return [torch.rand(batch_size, in_features)]
+    return [torch.rand(batch_size, in_features, dtype=torch.float16)]
 
 def get_init_inputs():
     return [in_features, out_features]
