+++ KernelBench/level2/84_Gemm_BatchNorm_Scaling_Softmax.py
@@ -9,7 +9,7 @@ class Model(nn.Module):
         super(Model, self).__init__()
         self.gemm = nn.Linear(in_features, out_features)
         self.bn = nn.BatchNorm1d(out_features, eps=bn_eps, momentum=bn_momentum)
-        self.scale = nn.Parameter(torch.ones(scale_shape))
+        self.scale = nn.Parameter(torch.ones(scale_shape, dtype=torch.float16))
         self.softmax = nn.Softmax(dim=1)
 
     def forward(self, x):
@@ -33,7 +33,7 @@ bn_momentum = 0.1
 scale_shape = (1,)
 
 def get_inputs():
-    return [torch.rand(batch_size, in_features)]
+    return [torch.rand(batch_size, in_features, dtype=torch.float16)]
 
 def get_init_inputs():
     return [in_features, out_features, bn_eps, bn_momentum, scale_shape]
