+++ KernelBench/level2/99_Matmul_GELU_Softmax.py
@@ -20,7 +20,7 @@ in_features = 8192
 out_features = 8192
 
 def get_inputs():
-    return [torch.rand(batch_size, in_features)]
+    return [torch.rand(batch_size, in_features, dtype=torch.float16)]
 
 def get_init_inputs():
     return [in_features, out_features]
