+++ KernelBench/level2/9_Matmul_Subtract_Multiply_ReLU.py
@@ -25,7 +25,7 @@ subtract_value = 2.0
 multiply_value = 1.5
 
 def get_inputs():
-    return [torch.rand(batch_size, in_features)]
+    return [torch.rand(batch_size, in_features, dtype=torch.float16)]
 
 def get_init_inputs():
     return [in_features, out_features, subtract_value, multiply_value]
